ch(refactor-serializers-and-views): refactor code
- refactor serializers
- refactor views
