+++ cars_site/cars/serializers.py
@@ -1,15 +1,6 @@
 from rest_framework import serializers
-from cars_site.cars.models import Car
+from .models import Car
 from cars_site.company.models import Company
-
-class CompanySerializer(serializers.ModelSerializer):
-    """Serializer to map the Model instance into JSON format."""
-
-    class Meta:
-        """Meta class to map serializer's fields with the model fields."""
-        model = Company
-        fields = ('id', 'name', 'location', 'image', 'established_year', 'rating')
-        read_only_fields = ('',)
 
 
 class CarSerializer(serializers.ModelSerializer):
@@ -19,7 +10,7 @@ class CarSerializer(serializers.ModelSerializer):
     class Meta:
         """Meta class to map serializer's fields with the model fields."""
         model = Car
-        fields = ('id', 'name', 'kind', 'year_of_creation', 'milleage', 'color', 'car_maker', 'is_available', 'car_engine_no', 'deleted', 'status', 'price', 'company_id')
+        fields = ('id', 'name', 'kind', 'year_of_creation', 'milleage', 'color', 'is_available', 'car_engine_no', 'deleted', 'status', 'price', 'company_id')
         read_only_fields = ('',)
 
     def create(self, validated_data):
