+++ cars_site/company/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+from .models import Company
+
+class CompanySerializer(serializers.ModelSerializer):
+    """Serializer to map the Model instance into JSON format."""
+
+    class Meta:
+        """Meta class to map serializer's fields with the model fields."""
+        model = Company
+        fields = ('id', 'name', 'location', 'image', 'established_year', 'rating')
+        read_only_fields = ('',)
